raspi-sports-classifier/pi3_consumer_processor/requirements.txt
kafka-python
pymysql
opencv-python
